Gracefully quit app when going into background on Android
Before this commit, touchHLE had no handling for going into the
background on Android, so audio would keep playing and the app had to be
manually killed, with loss of data. Now, it gracefully quits when going
into the background, just like when the window is closed on a desktop
OS. This is not the ideal, but it's a big improvement.

Also, `applicationWillResignActive:` is now sent when quitting, not just
`applicationWillTerminate:`.